chore(code style): automated code style clean up (#69)
* code cleanup

* fix tabs

+++ src/main/java/org/maxgamer/rs/core/server/TickableWrapper.java
@@ -32,7 +32,7 @@ public class TickableWrapper implements Comparable<TickableWrapper> {
 
     @Override
     public int compareTo(TickableWrapper t2) {
-        return (int) (this.getPeriod() - t2.getPeriod());
+        return this.getPeriod() - t2.getPeriod();
     }
 
     @Override
